requirements.txt
numpy==1.14.5
pandas==0.23.1
nilearn==0.5.0
scikit-learn==0.20.1
pybids==0.6.5
matplotlib==2.2.2
